move argparsing to utils
So that shell-completion can import it without causing a circular
dependency.
